refactor(bolt): move NewTestKVStore func to kv_test.go
